Add "DIFF" variable for latexdiff
Using a new variable DIFF makes the Makefile be in a different mode:
when a Git revision is given, it takes a latexdiff between the current
LaTeX files respect to those in the given revision. If a range of two
Git revisions is given, it takes a latexdiff between them. Example:

  make DIFF=HEAD^
  make DIFF=HEAD~3..HEAD