Fix make dist: avoid including the incomplete tarball itself
In some environments "ls" captures the incomplete tarball file that is
being generated by the redirection of the output.

Note that the option "--exclude FILE" is not in POSIX (well, actually
"tar" is not in the current POSIX and replaced with pax), but both GNU
tar and BSD tar support this option.