feat: allow rules for LINTS
Similar to TESTS, now LINTS variable accepts rules in the Makefile.
The LaTeX source file name is given as $1.

Loops in the "lint" rule are unrolled.

When linting fails for a source file, the process doesn't exit
immediately, continues for other sources, and exits (with a non-zero
status code) after finishing all files.